requirements.txt
Django>=3.2,<4.0
gunicorn
whitenoise
google-generativeai
social-auth-app-django
django-widget-tweaks
